Set default user status to disabled and added user update endpoint
This commit sets the initial state of a user as disabled and introduces a new 'UserUpdate' class and a specific router for user updates. The 'disabled' field in UserDBModel now has the default boolean as 'True'. This change was made to ensure that new users go through an approval process before they can start dealing with the application. The commit further allows the user to update their 'fullname' and 'email'. An added HTTPException prevents the provision of an invalid email, guaranteeing the email validation.