Refactor JWT token error handling in APIs
This commit updates the error handling for JWT token validation in the Update User and Get User API endpoints. Specifically, we shifted from using JWTError to InvalidTokenError from the jwt.exceptions for more specificity. This adjustment is done to ensure proper recognition and handling of invalid or non-existing authentication tokens, thereby enhancing application security and resilience.